refactor: split moq_decoder.rs (4812 lines) into focused modules
Break the monolithic moq_decoder.rs into moq_decoder/ directory:
- mod.rs (524) — shared types, module declarations, tests
- desktop.rs (1654) — macOS/iOS VTDecompressionSession + FFmpeg fallback
- macos_vt.rs (1204) — VTDecompressionSession FFI wrapper
- linux_gst.rs (807) — GStreamer VA-API/DMABuf decoder
- android.rs (625) — MediaCodec HardwareBuffer decoder

Also update .gitignore for damus/, divine-mobile/, dev/, build artifacts.
